Custom commands in Projects context menu
