+++ package.json
@@ -129,6 +129,7 @@
   "devDependencies": {
     "@electron-toolkit/eslint-config-ts": "^3.1.0",
     "@electron-toolkit/tsconfig": "^2.0.0",
+    "@electron/rebuild": "^4.0.4",
     "@eslint/js": "^9.17.0",
     "@playwright/test": "^1.49.1",
     "@tailwindcss/vite": "^4.0.0",
